implement user authorization

+++ routes/campgrounds.js
@@ -1,22 +1,9 @@
 const express = require('express');
 const Campground = require('../models/campground');
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
-const { campgroundSchema } = require('../schemas');
-const { isLoggedIn } = require('../middleware');
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware');
 
 const router = express.Router();
-
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    
-    if(error){
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
 
 //get all campgrounds
 router.get('/', catchAsync(async (req, res) => {
@@ -32,7 +19,12 @@ router.get('/new', isLoggedIn, (req, res) => {
 //show specific campground details
 router.get('/:id', catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findById(id).populate('reviews');
+    const campground = await Campground.findById(id).populate({
+        path:'reviews',
+        populate: {
+            path: 'author'
+        }
+    }).populate('author');
 
     if(!campground){
         req.flash('error', 'Cannot find that campground :(')
@@ -45,6 +37,7 @@ router.get('/:id', catchAsync(async (req, res) => {
 //add new campground
 router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, next) => {
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
 
     req.flash('success', 'Succesfully added new campground!');
@@ -52,7 +45,7 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 }))
 
 //get form to update campground
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
     const campground = await Campground.findById(id);
 
@@ -65,20 +58,26 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
 }))
 
 //update campground
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params;
-    const { campground } = req.body;
-    await Campground.findByIdAndUpdate(id, campground);
+    await Campground.findByIdAndUpdate(id, req.body.campground);
     req.flash('success', 'Succesfully updated campground!')
     res.redirect(`/campgrounds/${id}`);
 }))
 
 //delete campground
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+
+    if(!campground){
+        req.flash('error', 'Cannot find that campground :(')
+        return res.redirect('/campgrounds');
+    }
+
     await Campground.findByIdAndDelete(id);
-    req.flash('success', 'Succesfully deleted campground!');
-    res.redirect('/campgrounds');
+    req.flash('success', 'Succesfully deleted campground!')
+    res.redirect(`/campgrounds`);
 }))
 
-module.exports = router;
+module.exports = router;
